Return error when reading Parquet column with invalid utf-8 bytes
This PR switches the Parquet reading internals to return an error
instead of panicking when data with invalid utf-8 bytes is read. Per the
advice in #804 I did this by introducing new `try_*` variants of methods
which panic, and switched the callstack over. Thus, I believe this is an
API compatible change (`TraitBinaryArray` is changed incompatibly, but
it appears to be crate private).

I've attached a very small .parquet file with an invalid utf-8 byte that
I used for testing. Without this change using the
`parquet_read_parallel` example binary with the file panics, whereas now
it errors. I think this file should be added to the test suite, but I
need some guidance as to exactly how to structure that.