Add Json serialization of DateTime with timezone
Serializes DateTime types to the RFC3339/ISO-8601 format
(e.g. 1996-12-19T16:39:57+08:00).

DateTime serializng is done differently from serializing NaiveDateTime,
which uses a space instead of a T as the date and time separator.

However the default chrono to_string on timezone aware DateTIme produces
a space before the tz offset (e.g. 1996-12-19 16:39:57 +08:00)
which Javascript engines won't parse as a Date(), so the RFC3339 format
is used instead.

Serializing panics if invalid timezones are encontered, but this
improves on the todo!("still have to implement timezone") in
src/io/json/write/serialize.rs